It is time to push into master (#13)
* update-crateio-push-github-flow: updated push to crates io in order t… (#4)

* updated push to crates io in order to eliminate updating packages after pushing to develop

* Update crateio push GitHub flow (#5)

* Update crateio push GitHub flow (#7)

* updated the other github actions to trigger only after pushing or making pull requests into master branch

* Update crateio push GitHub flow (#8)

* updated build.yml

* Update crateio push GitHub flow (#9)

* added license into Cargo.toml

* NS-236/documentation: added README.md (#11)

* added README.md, described project

* Ns 276/provide logs for users (#12)

* updated log4rs

* updated log level in log4rs, addec ctrc handler, updated overall code structure

* added infinite capturing

* added table for displaying loaded config

* updated version to 0.1.1

+++ src/args.rs
@@ -6,7 +6,7 @@ pub struct Cli {
     #[arg(short = 'D', long, conflicts_with_all = ["config_file"])]
     pub device_name: Option<String>,
     #[arg(short = 'N', long, conflicts_with_all = ["config_file"])]
-    pub number_packages: Option<i32>,
+    pub number_packages: Option<u64>,
     #[arg(short = 'B', long, conflicts_with_all = ["config_file"])]
     pub buffer_size: Option<i32>,
     #[arg(short = 'C', long, conflicts_with_all = ["device_name", "number_packages", "buffer_size"])]
